divide class and long logic be changed to function
